% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{add_noise}
\alias{add_noise}
\title{Add noise to one matrix.}
\usage{
add_noise(data, mean = 0, sd = 0.2, val_min = 0, val_max = 1)
}
\arguments{
\item{data}{The matrix to disturb.}

\item{mean}{The mean of the normal distribution of noises.}

\item{sd}{The standard deviation of the normal distribution of noises.}

\item{val_min, }{the minimum value of the matrix}

\item{val_max, }{the maximal value of the matrix}
}
\value{
This function return the matrix with noise.
}
\description{
Add noise to one matrix.
}
