% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{age_effect}
\alias{age_effect}
\title{Function to compute A}
\usage{
age_effect(sites, effect)
}
\arguments{
\item{sites}{The names of probes affected by the age effect.}

\item{effect}{The list of linear regression between probes affected by the age and the age.}

\item{alpha0}{The variance between the patients, 1 is an high variance and 10,000 is a very low variance.}
}
\value{
This function a matrix of Dirichlet distribution.
}
\description{
This function make a Dirichlet distribution
}
