% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_treatments.R
\name{CF_detection}
\alias{CF_detection}
\title{Function to remove correlated probes}
\usage{
CF_detection(D, exp_grp)
}
\arguments{
\item{D}{The matrix patients*probes.}

\item{exp_grp}{The matrix of experimental data for all the patients.}
}
\value{
This function return the D matrix without correlated probes.
}
\description{
This function searchs the correlated probes by linear model.
}
\examples{

}
