% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{compare_A}
\alias{compare_A}
\title{Compare the A matrices}
\usage{
compare_A(A_r, A_est)
}
\arguments{
\item{A_r}{The real A matrix used for the simulation..}

\item{A_est}{The calculated A matrix.
Permutation is guessed from the minimum MAE.}
}
\value{
This function return the matrix with noise.
}
\description{
This function compute errors between the two A matrices,
  after guessing the permutation from the minimum MAE.
}
