% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{compute_A}
\alias{compute_A}
\title{Function to compute A}
\usage{
compute_A(n = 100, prop = c(0.3, 0.6, 0.1), alpha0 = 1)
}
\arguments{
\item{n}{The number of patients to simulate.}

\item{prop}{The different proportion of cell types in a vector of size k (number of cell types).}

\item{alpha0}{The variance between the patients, 1 is an high variance and 10,000 is a very low variance.}
}
\value{
This function return a matrix of size k*n with the cell type proportions.
}
\description{
This function make a Dirichlet distribution
}
