% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{compute_D}
\alias{compute_D}
\title{Computation of the D matrix.}
\usage{
compute_D(A_mat, T_mat, exp_grp, sites_sex = NA, coeff_sex = NA,
  plate_effect = NA)
}
\arguments{
\item{A_mat}{The A matrix.}

\item{T_mat}{The T matrix.}

\item{sites_sex}{NA by default, set sites affected by sex if you want a sex effect}

\item{coeff_sex}{NA by default, set coeff of sex if you want a sex effect.}

\item{plate_effect, }{Na by default, put plate_effect if you want one}
}
\value{
This function return a matrix of size probes*n with possibly an effect of confounding factors
}
\description{
This function compute the D matrix with an effect on age and / or plates
}
