% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{compute_TF}
\alias{compute_TF}
\title{Function to compute the female version of the T matrix}
\usage{
compute_TF(sites, coeff, T_brut)
}
\arguments{
\item{sites}{The names of probes affected by the sex effect.}

\item{coeff}{The coefficient of variation for probes affected by the sex.}

\item{T_brut}{The original T matrix.}
}
\value{
This function return a matrix modified for the sex.
}
\description{
This function makes a T matrix with probes modified by sex
}
