% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edec.R
\name{Edec}
\alias{Edec}
\title{Function to run Edec.}
\usage{
Edec(D, nbcell = 5, infloci)
}
\arguments{
\item{D}{The matrix patients*probes.}

\item{nbcell}{The number of cell types}

\item{infloci}{The list of informative probes.}
}
\value{
This function return a list with the A and the T matrices computed.
}
\description{
This function run Edec. and select A and T.
}
\examples{

}
