% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{MAE}
\alias{MAE}
\title{Computing the Mean Absolute Error}
\usage{
MAE(M1, M2)
}
\arguments{
\item{M1}{First matrix.}

\item{M2}{Second matrix}
}
\value{
This function return the mean absolute error between the 2 matrices.
}
\description{
This function compute the mean absolute error between 2 matrices
}
