% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDC.R
\name{MDC}
\alias{MDC}
\title{Function to run MeDeCom.}
\usage{
MDC(D, nbcell = 5, lambdas = c(0, 10^(-5:-1)))
}
\arguments{
\item{D}{The matrix patients*probes.}

\item{nbcell}{The number of cell types.}

\item{lambdas}{Values of parameter lambda to be tested.}
}
\value{
This function return a list with the A and the T matrices computed.
}
\description{
This function run MeDeCom and select A and T.
}
\examples{

}
