% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medepir-package.R
\docType{package}
\name{medepir-package}
\alias{medepir}
\alias{medepir-package}
\title{medepir: MEthylation DEconvoluation PIpeline in R}
\description{
A pipeline to quantify cell-type heterogeneity from DNA methylation 
    data. Package medepir provides practical implementation of recommended 
    guidelines for inference of cell-type proportions with confounders, based on 
    non-negative matrix factorization of DNA methylation data.
}
\author{
\strong{Maintainer}: Magali Richard \email{magali.richard@univ-grenoble-alpes.fr}

Authors:
\itemize{
  \item Clémentine Decamps \email{clementine.decamps@gmail.com}
  \item Florian Privé \email{florian.prive.21@gmail.com}
  \item Michael Blum \email{michael.blum@univ-grenoble-alpes.fr}
}

}
\keyword{internal}
