% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_treatments.R
\name{plot_k}
\alias{plot_k}
\title{Plot the ACP to choose k}
\usage{
plot_k(D)
}
\arguments{
\item{D}{The matrix patients*probes.}
}
\value{
This function returns the plot of the ACP.
}
\description{
This function makes a PCA to search the number k of cell types.
}
