% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_treatments.R
\name{pval_CF}
\alias{pval_CF}
\title{Function to compute the p-value of the CF}
\usage{
pval_CF(val, labels)
}
\arguments{
\item{val}{expression}

\item{labels}{counfounding factors}
}
\value{
this function return pvalues
}
\description{
This function makes a linear regression and compute the pvalue
}
