% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medepir-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{nb_cores}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bigstatsr}{\code{\link[bigstatsr]{nb_cores}}}
}}

