% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFE.R
\name{RFE}
\alias{RFE}
\title{Function to run RefFreeEwas with euclidean dist.}
\usage{
RFE(D, nbcell = 5)
}
\arguments{
\item{D}{The matrix patients*probes.}

\item{nbcell}{The number of cell types}
}
\value{
This function return a list with the A and the T matrices computed.
}
\description{
This function run RefFreeCellMix and select A and T.
}
