% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_treatments.R
\name{feature_selection}
\alias{feature_selection}
\title{Function to select probes with a high variance}
\usage{
feature_selection(D, varmin = 0.02)
}
\arguments{
\item{D}{The matrix of dimension patients*probes.}

\item{varmin}{The minimum variance to be selected.}
}
\value{
This function return the D matrix with selected probes.
}
\description{
This function selects probes with a variance > varmin
}
